chore: move data-provider directory to botonic-server
