+++ packages/botonic-core/package.json
@@ -30,15 +30,11 @@
     "README.md"
   ],
   "dependencies": {
-    "aws-sdk": "^2.997.0",
     "axios": "^0.24.0",
     "decode": "^0.3.0",
-    "dynamodb-toolbox": "^0.3.4",
     "fast-xml-parser": "^3.19.0",
     "html-entities": "^2.3.2",
-    "node-json-db": "^1.3.0",
-    "pusher-js": "^5.1.1",
-    "ulid": "^2.3.0"
+    "pusher-js": "^5.1.1"
   },
   "devDependencies": {
     "@babel/cli": "^7.11.5",
