--- packages/botonic-core/src/data-provider/sqs-utils/hashing.ts
@@ -1,5 +1,0 @@
-import crypto from 'crypto'
-
-export function sha256(message: string): string {
-  return crypto.createHash('sha256').update(message).digest('hex')
-}
