--- packages/botonic-core/src/data-provider/sqs-utils/index.ts
@@ -1,2 +1,0 @@
-export { enqueueToSQS } from './queue-to-sqs-decorator'
-export { buildSendMessageRequestForQueue } from './send-message-request'
