+++ packages/botonic-server/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "@botonic/api",
-  "version": "1.0.0-dev.1",
+  "name": "@botonic/server",
+  "version": "1.0.0-dev.0",
   "description": "",
   "main": "src/index.ts",
   "scripts": {
@@ -28,16 +28,18 @@
     "@botonic/core": "dev",
     "@vendia/serverless-express": "^4.3.9",
     "aws-lambda": "^1.0.6",
-    "aws-sdk": "^2.940.0",
+    "aws-sdk": "^2.1023.0",
     "axios": "^0.24.0",
     "bufferutil": "^4.0.3",
     "chalk": "^4.1.1",
     "cors": "^2.8.5",
+    "dynamodb-toolbox": "^0.3.5",
     "express": "^4.17.1",
     "express-jwt": "^6.1.0",
     "express-validator": "^6.12.0",
     "express-ws": "^5.0.2",
     "jsonwebtoken": "^8.5.1",
+    "node-json-db": "^1.4.1",
     "terminal-link": "^3.0.0",
     "ulid": "^2.3.0",
     "utf-8-validate": "^5.0.5",
