+++ packages/botonic-server/src/constants.ts
@@ -1,0 +1,6 @@
+export enum Environments {
+  AWS = 'aws',
+  LOCAL = 'local',
+}
+
+export const WSS_PROTOCOL_PREFIX = 'wss://'
