+++ packages/botonic-server/src/data-provider/dynamodb-utils.ts
@@ -1,8 +1,7 @@
+import { MessageEventTypes } from '@botonic/core'
 import { DynamoDB } from 'aws-sdk'
 import { Entity, Table } from 'dynamodb-toolbox'
 import { EntityAttributes } from 'dynamodb-toolbox/dist/classes/Entity'
-
-import { MessageEventTypes } from '../models'
 
 // Table Definitions
 export const GLOBAL_SECONDARY_INDEX_NAME = 'GSI1'
