+++ packages/botonic-server/src/handlers/index.ts
@@ -1,3 +1,2 @@
 export * from './bot-executor'
-export * from './dispatchers'
 export * from './sender'
