+++ packages/botonic-server/src/notifying/index.ts
@@ -1,0 +1,1 @@
+export { enqueueToSQS } from './queue-to-sqs-decorator'
