+++ packages/create-botonic-app/dev-template/api/src/handlers/botExecutor.js
@@ -1,7 +1,7 @@
 // eslint-disable-next-line filenames/match-regex
-import { dispatchersFactory } from '@botonic/api/src/dispatchers'
-import { botExecutorHandlerFactory } from '@botonic/api/src/handlers/bot-executor'
-import { dataProviderFactory } from '@botonic/core/lib/esm/data-provider'
+import { dataProviderFactory } from '@botonic/server'
+import { dispatchersFactory } from '@botonic/server/src/dispatchers'
+import { botExecutorHandlerFactory } from '@botonic/server/src/handlers/bot-executor'
 import { app as bot } from 'bot'
 
 import sender from './sender'
